Update for NLTK >=3
NLTK now supports Python 3, so this commit adds the option to use NLTK's
Tree class directly as well as updating the internal tree_mini module,
which is a detached version of the NLTK Tree class.  Because of
weirdness with subprocesses, use of full NLTK is off by default, but can
be triggered by an option to reload_lamb (or a global flag in
lamb.utils).