Add Drew's compositional DRT notebook
This notebook implements a piece of compositional DRT, with some advanced
techniques for modifying the lambda notebook.  I've modified this slightly
from Drew's original version given changes to the lambda notebook in the
interim.

This commit also forces a change in behavior for assignments; they are now
assumed to be never modified by side-effect.  I _think_ this assumption has
been in play for some time in TypedTerm, but I'm not entirely confident.