Add flexibility to .latex_str() with kwargs
This change also comes with a default to not display types for
bound variables except on the binder.  This should make complex
expressions much more readable, at the expense of explicitness.
There is a global preference in `meta` to change this.  This does
not change the results of `repr`, which is explicit about every
type.