Defer partiality calculation until after beta reduction
This drastically simplifies and clarifies the results -- repeated
instances of beta reduction and calculate_partiality were leading to
fairly weird conditions, when the argument to a lambda expression
contained a Partial.  In principle this seems like it should be an
avoidable problem, but I don't really see how to avoid it now.