Document an issue with PA and partiality
Because PA introduces a lambda term after partiality has been calculated
for the content node, the content node can't know to defer calculation
of partiality until beta reduction has been done. Consequently,
partiality conditions are over-generated if the argument to a
PA-introduced lambda has a Partial. Not immediately sure what the right
solution to this is (beyond having a proof system).