Rework partiality, again
This reworks partiality simplification for quantified expressions to be
better grounded in a model theory I've worked out (though it's not
discussed in detail in the documentation). There are two projection
modes, weak and strong, with weak being the default: this imposes the
most minimal definedness condition possible to interpret the body of the
quantified expression on a model. Strong requires the whole model to
obey the definedness condition.  The first version was very ad-hoc, and
on reexamination, probably wrong.

As part of this, I changed the basic Iota to not introduce definedness
conditions, splitting the previous behavior off into `IotaPartial`.
Simplifying IotaPartial will only generate regular Iotas.

This commit also fixes a number of bugs in partiality simplification
that were over-complicating some derivations, and ensures that
`derivation` is set on calculating partiality, and probably a few other
cleanups.