Rewrite derivation display code
The previous display code was overly complicated, hard to modify (in
part because of the way templating was handled), and relied heavily on
tables, which are not generally reliable when embedded in the output
cell of a jupyter notebook. The dependence on tables was leading to at
least two problems: (i) Jupyter's table styles, really intended for
tabular data of the numpy vein, were being inherited and overriding them
(which I had tried to do in the past) was a moving target. (ii) MathJax
in recursive tables was not calculating widths correctly, leading to bad
(sometimes very bad) rendering in HTML-CSS mode, the default in Jupyter,
with lots of spurious linebreaks inserted.  In addition, though it
sometimes resembles it in limiting cases, the things I am displaying
with this code are not really tabular data, so modern web design
principles would say not to use <table>.

This rewrite makes three main changes. First, it separates out the
structure of a recursive derivation display from the styling (one of the
problematic elements of the previous implementation that made it hard to
work with). Second, it does all html templating with
xml.etree.ElementTree, rather than generating html directly. This does
seem to result in slightly uglier code, but still code that is *much*
more tractable to modify later.  Third, it moves to all divs (with just
one exception where a derivation is displayed as purely tabular). I had
tried to do this before once, but at the time the existing display code
proved too hard to adapt -- eventually leading to this rewrite.

This commit fixes all the problems with MathJax/HTML-CSS rendering and
linebreaking that had developed, as they result from the bad
interactions of MathJax and recursive tables.