Work around various differences between firefox vs chrome
This have to do with basically three things: differences in how strict
the browser is about div-based tables that e.g. don't have the same
number of cells in each row; differences in when line-breaking happens;
and differences in how sizing of divs under certain embedded conditions
are handled (leading to mathjax linebreaking, again).

Pretty much everything should now look ok in both firefox and chrome,
for both html-css and svg renderers, in a live notebook. There is still
spurious linebreaking happening in html-css in nbconvert outputs, as
well as spurious resizing for svg mode. (Common HTML looks great in
nbviewer outpus, but this isn't even working for me in current jupyter
lab.)