Work around yet another mathjax line break oddity
If a mathjax formula is in a table row with a non-mathjax span in
another row, and the other span is narrower, then in HTML-CSS rendering
mode the mathjax span will do some linebreaking to try to match the
narrower row. Wrapping the mathjax in an inline-block div apparently
prevents this.  As a consequence I'm just going to wrap all text in an
inline-block div, which is maybe overkill...

The div structure is growing increasingly complicated in order to deal
with these linebreaking issues, but at the moment it all seems
like necessary complexity.