Use IPython.display.display where appropriate
`display` is much better than the custom function I was using for
similar effect from several years ago. I've also changed it so that
lamb.display (and lamb.parsing) is not automatically part of the
environment, so it doesn't shadow `display`, which is nowadays set up
without needing an import. (Though I have added import statements in
order to be explicit in some cases.)