Various improvements in how composition failures work
This commit has one big technical change and many jupyter display
improvments for cases where composition fails. The technical change is
that now, composition failure will only use a TypeMismatch when that is
actually appropriate: a new exception CompositionFailure is for the more
general case of composition failing for any reason. A subclass handles
cases where the preconditions fail, and these are usually only shown in
a compressed format. The display improvements focus on cases where
composition involves parts that themselves have failed to compose;
errors in these cases should now be much more informative, and the api
is a little bit cleaner.

This commit brought to you by jury duty.