fix: explicit length check when parsing binding strucs
Rather than raising an exception (which leads to a cryptic parsing
error), return None. This leads to an exception being raised later, on
eval("") more or less. In particular, this causes `%te` without an
argument to raise an unexpected EOF exception. This might not really be
any better from the user perspective, but is at least now happening
intentionally.