refactor: further clean up kernel installation
This deals with several issues I've been banging my head against, where
the main one is allowing the lambda notebook kernel to be installed in
the current env rather than per-user. This also makes it easier to have
a "-dev" kernel installed which is using code from a repo, at the same
time as a regular lambda notebook kernel that is using a system install.