install: include kernel directly in bdist
Previously, the user needed to run an install_kernel script or function.
However, if the lamb package is installed, there's nothing complicated
or custom about the kernelspec that needs install-time programmatic
generation, and so we can just install the kernelspec directly. This was
a bit tricky to figure out how to do in the current python installer
land, and I probably am doing it "wrong" by even using setup.py, but if
it is wrong, then so are a lot of other kernel packages up to and
including ipykernel (which was my starting point for the setup code).

(Commit still needs some testing probably...)