feat: improve binder behavior
Previously a Binder took its index from its form, rather than its index
method. I think this was originally because of weird compatibility stuff
with nltk.Tree, which I've now fixed. This allows Binders to have a form
beyond the number.

Also, minor improvements in wording in the relative clause example. This
notebook is not backwards compatible before the change in this commit.