feat: improve support for vacuous elements
Binders were implemented as vacuous elements with an index, by way of
setting content to `None`. This commit improves support for vacuous
elements beyond this, by (a) fixing some cases where they would lead to
unwarrented crashes, and (b) providing a composition operation (for tree
composition in particular) that explicitly ignores them.

Also, adjust the relative clause notebook to use vacuous elements for
"such" and "that in the such-that RC example.