refactor: exercise more fine-grained control over reprs
The original approach (which is nearly as old as this project) to reprs
is pretty brute force: essentially everything is displayed via
_repr_html_ because in standard Jupyter, this works well with mathjax.
However, this causes a bunch of issues:
* In colab, _repr_html_ does not render latex
* at the command line, html code was displayed in a bunch of cases where
  it shouldn't be
* In some cases, text repr isn't available if the frontend wants that,
  even though the code was there
* In some cases _repr_html_ was used for pure latex output

This commit provides some scaffolding for finer-grained control over
what repr is used where, and fixes a bunch of cases. In some cases,
_repr_markdown_ is used, since in colab, this leads to latex rendering
and still allows some html (generally not styling, in my testing). Text
reprs are handled a little more cleanly for ipython output. This is all
still a little bit piecemeal, so standardizing some of this api could be
very helpful. It also doesn't deal with the biggest problem, i.e.
derivation trees.