fix: stabilize order in recursive polymorphic unify
This fixes what was apparently a regression in 70102933889 (!). The
order in which non-polymorphic types are unified was accidentally
swapped in that commit, leading to inconsistencies. No wrong answers
were actually returned -- these are all cases where the principal type
could be expressed in multiple ways. But, `meta` code for doing type
specialization relied on a stable order here, so this led to bugs with
complex polymorphic combinators (noticeable in the Hamblin fragment at
least). The current solution is probably more general.

This commit also better documents what's going on, and adds some
tests for this case.