feat: prototype colab html+latex rendering
This commit (finally) accomplishes working rendering for the parts of
lambda notebook that embed latex math code within a complex html tree.
See the comment added in this commit for details, but the challenge is
that colab renders outputs in an iframe and uses katex (not mathjax) in
a fairly idiosyncratic way when doing so. It doesn't call katex at all
on HTML outputs, markdown outputs strip html attributes, and HTMLMath
widgets simply don't work; so I had a challenge figuring out how to get
any sort of rendering at all to happen for these complex outputs. This
commit injects katex autorender to lambda notebook HTML outputs, and
calls it.

I also took this opportunity to rename MiniLatex, since it's no longer
very descriptive of what it does.