fix: improve handling if type variable binding
This commit fixes a number of issues that amount to type variable
binding not being handled correctly for application expressions. The fix
in 99c115676 was broken for the argument position, which is good,
because the intended fix was completely wrong. Instead of compacting the
function and argument independently, we need to compact them as a whole,
pushing the variable binding up a level. There were also some issues
with how `let` was percolated in reduction, as well as how
function-argument history was tracked.

Also, update the type variable docs a bit.