fix: remove `reload_lamb()` from most distributed notebooks
This has been here because it is very convenient for me while testing
things. However, beause of the weirdness of stale objects that results
from running it out of sequence, I don't want to advertise it as if it
is something that new users should be calling a lot. Restarting the
kernel accomplishes the same thing as calling this at the top of a
notebook anyways.

It still sticks around for a few special cases, and I've added various
caveats to some of this.

Also, a few misc changes to various notebooks, mostly wording /
stylistic changes.