fix: latex rendering compatibility
We now have potentially 3 web latex rendering targets (Mathjax 2/3,
Katex), with slightly different compatibility among them. As suggested
by the commit message in 900d96ef3, it does turn out that MathJax 3's
bold behavior is not backwards compatible. There are also issues with
`_` in text. This commit implements modes for each of the three as well
as a compatibility mode that is intended to work everywhere.
Unfortunately, I can't figure out how to get text bold to work in
compatibility mode.

(Side note: I think that MathJax 3 is probably the most faithful to
actual LaTeX.)

Currently, KATEX mode is determined by whether colab is running, and
otherwise COMPAT mode is the default.  Perhaps in the future I can find
a way of actually setting the renderer mode based on some system
features (jupyter version + the existence of the mathjax 2 extension?).