fix: better latex mode defaults
Try to choose between mathjax 2/3 in a smarter way.  This is still not a
fully complete implementation, as it assumes that if
`jupyterlab_mathjax2` is installed, it is enabled. But this is not a
terrible assumption, I suspect. Also, it's possible this would break on
an earlier version that had the mathjax 3 extension installed? (If I
understand the previous situation.)

The katex extension is just completely broken for my purposes, so I'm
not going to support it.