feat: improve handling of type parsing errors, documentation
This commit has a bunch of stuff in it, but it's generally about how
types are documented, especially around how errors are handled. The
commit adds a line magic for type parsing (`%tp`), and refactors how
`TypeParseError`s are used. As part of this, it adds better repr for
these errors. Ultimately, some of the `%te` code that logs errors
could also be converted along these lines. Specific notes:

* Previously, `tp` as used in demos and docs would not error if there were
  characters past the end of the type; this is because this is the correct
  behavior for metalanguage parsing. The convenience function and line
  magic now do error for this case.
* Many documentation tweaks that happened while updating docs to use
  `%tp`.
* Disjunction type docs had a bunch of unrelated issues, mainly to do
  with how functions were discussed. I have removed the
  `factor_functional_types` cases entirely, though left the function, as
  I realized this can actually change the type. The function may still
  be useful.
* Parsing issues with disjunctive types
* General improvements to metalanguage type docs