fix: flag that VSCode HTML+latex is broken
This package relies on MathJax within HTML working in order to render
derivations, and this is broken in VSCode (the relevant bug is
https://github.com/microsoft/vscode-jupyter/issues/7801, but it seems to
have been auto-closed without resolution; see also
https://github.com/microsoft/vscode-jupyter/issues/12537 and maybe
https://github.com/microsoft/vscode-jupyter/issues/8131). It's possible
that given time I'll be able to find a workaround (as I did for colab),
but this looks if anything less trivial than colab -- since there's no
browser tools etc., and google searches are much less promising. For
now, directly flag the issue to users as a vscode-jupyter issue.