feat: preserve compose order in bottom-up composition
Bottom-up composition has a very strict partitioning between object
language and composition operations, where the operation can't even see
the object language word-order. This led (a very long time ago) to a
weird quirk of derivations in this approach: F-A order is used in trees
and tree-like representations. This commit changes things to use the
input word order for ordering in derivation trees. It does so by adding
yet another convoluted step on top of an already convoluted system; the
tree order is swapped back for reverse application. Ultimately, most of
this code could probably use some simplification -- I would not write it
now the way I did 10 years ago.