feat: function to convert composed trees to plain `Tree`s
These objects inherit from Tree, but actually calling the base Tree's
rich repr code is pretty painful (at least for TreeComposite), because
we put things other than strings in the label slot. For CompositionTree
this rendering code still works, so there would be other ways of simply
displaying the underlying structure, but it still seems useful to get to
the plain tree based on the stored node labels, and also to have a
consisten api.

Differences between the two:
* in bottom-up composition, node labels are built simply from parts.
  This is true of bottom-up composition in both systems, but for
  `CompositionTree`s, if there are provided node labels, they will be
  used instead.
* TreeComposite bottom-up composition will project extra non-branching
  nodes for type-shifts
* CompositionTree represents multiple derivations with a single source
  tree.

As a convenience aimed at quick display, I've also added a version of
this function on CompositionResult; it will return a list of `Tree`s if
the size of the results is not 1. This matches CompositionTree (which
always returns a single tree) for this case.