feat: (re)-add some heuristic simplification rules
This is a partial revert of a removal in 05b1fa331d54. I removed these
rules because it seemed extremely ad hoc (and ultimately pointless) to
have simplification rules that worked by syntactic equivalence. This is
no less true, but these cases do come up in practice, and formulas that
don't do at least this amount of simplification in practice can be
extremely annoying to deal with. I've also added excluded
middle/non-contradiction, as well as double negation.

All of this is still awaiting a much more robust approach to
simplification, e.g. using an actual inference engine, or at least some
kind of formula normalization. Alternatively, there's a literature on
boolean minimization that could be useful for these specific cases.