refactor: put metametalanguage on more solid ground
Prior to this commit, there were bits and pieces of a metametalanguage /
ontology, but they were very incomplete and very inconsistent. For types
e and n, python True/False and numbers could be used to construct terms
that effectively acted like references to elements of the corresponding
domain, but there was no explicit connection to this domain, and very ad
hoc handling of the connection between e.g. the term `True` and the
python object True. These terms also could be remapped under assignment
in a very odd way, and had some other quirks (because python 0/1 and
False/True hash together, and bool is a numeric type, a large number of
lurking bugs were present in formulas that mixed them). This commit
revamps all this completely to treat the metametalanguage / ontology
systematically, using type domains backed by underlying python objects.

* Type domains are fleshed out and actually used: the idea is that types
  should be backed by python objects, and these classes provide a
  rudimentary interface for actually doing this, e.g. checking
  membership. (Finally -- I think the stub classes this builds on were
  some of the first things I wrote 10+ years ago, and never returned
  to.)
* type_t is backed by {True,False} and type_n is backed by python
  numerics (really, ints, though this is a bit wishy-washy still).
* For non-t/n types, the default is for the type domain to consist
  of arbitrary strings involving a type-specific unique prefix followed
  by any numerals. For now, nothing much uses these elements, but they
  work in terms of parsing. The default prefix for type e is `c` (and
  prefixes `w`/`e` are more or less reserved for worlds and entities).
* A new term class, `MetaTerm`, is used to refer in a fully constant way
  to elements of type domains. Parsing and repr handling is essentially
  the same for n/t as it was before, though MetaTerms in rich reprs are
  now rendered in sans serif.  slightly different. To refer to elements of
  a string-backed type domain, the designated prefix `_` is used. Regular
  TypedTerm can no longer be used for domain elements and will error on
  manual construction.
* Noticeable tweaks to: unary positive/negative operators, tuples and
  tuple indexing, parsing of _, truthtables, to_python/from_python,
  true_term/false_term, random expression generation

This is all relatively undocumented for the moment as I continue to
tweak things related to this change.