fix: tweak parsing behavior for 0/1
This provides a casting-like behavior for MetaTerms during parsing only
between types t and n. Without explicit types, numbers give n and bools
give t, but after this commit, with explicit types, you get the expected
pythonic conversion between the two (e.g. `True_n` gives a reference to
`0`). Once instantited, the type distinction between the two remains
exact, and t and n are not interchangeable.

It's not actually clear to me whether this kind of parsing shortcut is a
great idea from the perspective of language principles, but it does
better reflect the underlying nature of the python types (e.g.  `bool`
is a subtype of `int`).