fix!: remove `lambda_notebook.py` etc
This project originates from a time where jupyter wasn't very common,
and I went to a lot of effort to make it easy to open lambda notebook
files for non-experts, including various attempts at a "branded" jupyter
invocation; this commit removes the last of these.

Jupyter is pervasive in data science and a lot easier to install and run
than 10 years ago; the lambda notebook is on pypi, and the kernel works
pretty well in jupyter lab generally. So, this all seems much less
necessary. In addition, this code is kind of brittle, and not entirely
easy to fix when it does break (it's broken on current jupyter). I no
longer use it myself, either. So, let's hope that people can figure out
how to run jupyter lab on their own in 2023, and remove it + its
maintenance burden.

The `lambda_console.py` script + supporting code stays, as it's
incredibly useful for development.

README.md needs a much bigger update, but this is the minimal change
necessary for the current state of things.