feat: expand out type domain elements
This commit provides type domains for all non-atomic non-variable types,
bringing the type domain set interface much closer to my goals. As with
the atomic type domain sets, the idea is to have elements correspond to
python objects, which can be wrapped in a `MetaTerm`. They are therefore
non-complex from the metalanguage perspective, but implement the main
features of the relevant types. Currently, none of the complex domain
elements can be instantiated via the metalanguage parser, but rather,
must be constructed via `MetaTerm` directly.

* A tuple element is backed by python `tuple`. These implement indexing.
* A set element is backed by python `set` (implementing only finite
  sets). These implement basic set membership.
* A function element is backed by python `dict` or python `set`, which
  is treated as a characteristic function on an entire domain. This
  provides a limited means of implementing non-finite functions; the
  dict approach corresponds to the standard model-theoretic way of
  defining functions. Note that a set will only be wrappable in a
  functional type with explicit typing or an explicit parameter. These
  implement reduction. In the future I may also add a wrapper for python
  functions. A dict-based function given an argument that is not in its
  domain will raise a specific new exception on reduction.
* Disjunctive types have a domain set but always reduce to their
  concrete types. (It could be useful to allow disjunctive types in
  MetaTerms in certain cases though; tbd.)

Recursive structure is possible as long as the elements are also in some
type domain in the current system. So for example,
`MetaTerm({'_c1', '_c2'})` constructs an element of the set
corresponding to type `{e}`, and with either an extra `typ=tp("<e,t>")`
or `setfun=True` would allow constructing the characteristic function
for that set.

Other misc notes and changes
* `MetaTerm` now generally compares equal to and hashes with its
  contained object.
* A fix to how sets are parsed
* tweaks to various aspects of atomic domain sets