feat: add {} to the parsed bracket list
This changes a bunch of internal handling of {} during parsing,
hopefully for the better; one thing it should guarantee is that sets are
always converted to TypedExpr before any operators are applied. The
prior commit fixed some aspects of this, but the parser would still
error on cases like `{x_e, y_e} <=> {x_e, y_e}`.