feat: term order normalization for many expression types
This commit implements an order based normalization for operators that
are associative and commutative, as well as for binding operators (which
are left commutative typically). The point of this is to reduce easy
sources of variance for both syntactic comparisons and operations that
rely on syntactic structure. For example, in commutative/associative
sequences, it puts syntactically identical terms adjacent to each other,
with a stable (left) associativity that matches the operator parsing
behavior; something like `q_t & p_t & q_t` previously would not
simplify, but now converts to `p_t & q_t`, which would result from any
combination of these terms conjoined in any order with any
parentheticalization.

Things that come along with this commit:
* refactoring of simplification/reduction/manipulation code (including
  derivations) into `meta.ply`
* generalization of simplification code for associative/commutative
* operators: after alphabetic normalization, this code now runs on
  adjacent expression pairs ensuring that all easy simplifications are
  applied to all potentially relevant pairs.
* Various improvements to parentheses in rich reprs
* improvements to derivation displays
* join api is changed to match python `join` and generalized a bit. It
  now has a counterpart function `collect` that feeds joinable types.
  BindingOps also have a `join` implementation.