fix: reduction improvements
The previous reduction code was extremely brute force, and had some
pathological behavior in the worst case. The general problem is that
individual reductions may create subexpressions that can newly be
reduced after the reduction, and can also generate an expression that
itself may need to be reduced. So in the worst case, a reduction
algorithm on t may need to reduce t some, then do recursive reductions,
then reduce t, etc. The previous code handle this just by recursing
(restarting reduce_all) if anything happened on the recursive step. This
can get a bad combinatorial blowup for pathological cases.[*]

However, the problem really doesn't require brute force. A full
reduction requires an invariant finite number of steps, and the trick is
simply finding and ordering them.

This code substantially improves by doing two things. First, on every
step in reduce_all, we do a top-down search (breadth first by default)
to find a next (sub)expression that needs to be reduced, instead of
blindly recursing. Second, this search is memoized, so that it builds a
chart of subexpressions that no longer need reduction (updating
appropriately on beta reduction, which is the only place this can change
non-monotonically during reduction).

I should add that I'm still a bit unclear on the worst case for this
algorithm. There's a literature on reductions that suggests (based on a
cursory read) that this code still likely has some pathological bad
cases[*] wrt how it uses copying; see e.g.:
https://dl.acm.org/doi/pdf/10.1145/96709.96711

[*] Probably, none of this has ever mattered for anything that anyone
uses the lambda notebook for.