fix: improve some low-level term handling code
This code for distinguishing symbol types was kind of bad, and also got
called a *lot* in the course of things like beta reduction. The
multi-word check also seems to have just been wrong, in that it accepted
strings designed to be stripped, but was not used in code that stripped
symbol names. (Maybe in the distant past I did use it in parsing?)
Anyways:

* Convert the low-level checks into much simpler regex-based checks
* Move into the TypedTerm constructor so that var vs constant is not
  recalculated on every call