feat: quantifier semantics for finite domains
This commit (more or less) completes first-order semantics for finite
domains, by implementing quantifiers.

* Important note: there is currently exactly one finite domain, for
  type t. (Future commits will address this.)

It implements quantifier semantics in two forms. Under a regular
simplify call, if the scope is completely evaluable (no free terms of
any kind, with a finite domain for the quantified type), it computes the
result in python code by iterating over the domain and performing
substitutions. This is brute force but should hopefully be fine for
relatively small domains. (There are various optimizations I can
imagine, e.g. looking for Horn clauses for universal quantifiers, but
then we do get to the inevitable issue that something fancier would be
better calling out to a theorem prover.) Second, non-default, there is a
simplify option that can cause quantified expressions with enumerable
domains to be converted to non-quantified boolean expressions. This can
blow up, but it's also quite useful for understanding what might be
going on here. This can also be triggered directly with an `eliminate()`
call on the object.

* Regular quantifier semantics: this doesn't track full derivations for
  the cases where a result is determined by arbitrary quantification
  over a domain (e.g. true universal quantification, false existential
  quantification)
* Iota and IotaPartial do implement regular quantifier semantics, but
  not any sort of elimination. (In principle, one could eliminate them
  by converting to widest scope ExistsExact?)
* Eliminate semantics: This is implemented straightforwardly for
  universal/existential to conjunction/disjunction, less
  straightforwardly for ExistsExact which first converts to a standard
  first order implementation, and also for ConditionSet, which converts
  to ListedSet. Elimination can be triggered directly by calling
  `eliminate()` on the object. The ConditionSet case only applies if the
  condition is fully evaluable, the others will work regardless.

Other subsidiary changes:

* bugfixes for using MetaTerms in assignments
* improved api for triggering pre-simplification
* some tweaks to ListedSet simplification, IotaPartial simplification
* Partial rendering now uses \xcancel for false conditions, which looks
  great in MathJax 3 (but I suspect won't work at all elsewhere?
  untested...) Also, some bugfixes/improvements to Partial rendering,
  derivations
* a refactor of how multisimplify handles the initial normalize step,
  with some consequent Derivation changes. (The current set of changes
  revealed some bugs with this code that I couldn't manage to fix any
  other way.)

fixup