feat: more compact notation for MetaTerms under assignment
When a MetaTerm is generated via an assignment under certain
circumstances, track the term's name and use that as the primary
notation for the MetaTerm in resulting derivations. This is done for all
constants, and for variables only when they are instantiated as part of
a quantifier derivation.  For atomic typed MetaTerms, the rendering
still shows the exact value (now in a superscript on the naming term),
but for more complex ones, it elides the value unless directly inspected
-- the idea being that functions and the like get very complicated very
quickly to show.  This notation may be more cryptic in some cases (it's
not particularly standard for the atomic typed case), but it is overall
an improvement I think.