feat: improvements to partiality
The implementation of partiality has been lingering for quite a while;
this commit provides yet another refactor. There are many improvements
to the calculate_partiality logic (it is much more correct now),
including integration into the simplification system. The biggest change
is to not automatically merge Partials in a condition into a single
flattened condition. In general, this kind of recursive partial should
be viewed as a sequence of definedness conditions. This commit also
implements operators that extract just the body or condition of a
Partial expression, and updates the documentation to provide a more
plausible account for the most complicated example there.