fix: revert some broken unification caching
Unification caching in 2672871a7f24 tried to be clever about
UnknownTypes, but this was really broken (I'm not sure what I was
thinking). First of all, this would cache pairs like <?1,?2> and <?1,?1>
together. Second of all, because the cache stores a full
UnificationResult, it would have needed to update the mappings when
returning a cached result based on non-exact UnknownType equality. This
is complicated and probably not worth it. For now I've just reverted any
attempts at cleverness, though this probably undoes some of the apparent
speedups from 2672871a7f24. I've left some of the api though.

Other changes that happened while finding this:

* debug printing infrastructure for managing indents when output
  debugging recursive calls
* refactor of type adjustment cache (done while looking for side effects
  generally)
* other misc tweaks