fix: change handling of scope assignment type checking
Previously, this seems to have been intended to use the entire
assignment for type checking. But, ensure_typed_expr actually carries
out variable replacement if it's provided an assignment, and this isn't
what should happen on simply constructing a BindingOp. Instead, just use
an assignment containing only the bound variable. It is (hopefully) the
case that type checking from the assignment will still happen for any
variables.