refactor: generalize anonymous let-polymorphism
Previously, handling of `let` was very ad hoc and it only even worked on
functions and application expressions; it was only really correct for
application expressions and then only under certain assumptions about
the context. This happened to be most of the cases I was using it with,
and this was enough for (anonymous) combinators to work as
let-polymorphic composition operations.

However, it was extremely non-general, with all sorts of weird bugs
lurking. This especially has interfered with getting even the most basic
variable substitution working with type variables. This commit
generalizes so that `let` is handled correctly in every instance of a
`factory` call.  The result is much cleaner behavior. This also allowed
some tweaks for key special cases, e.g. if all subelements have `let`
set, then so will the result -- there's no need to freshen every type
variable. Type variable compaction also now only targets fresh types.