feat: an explicit type constructor for variable binding
This commit implements an unselective type variable binder, with
associated unification code. (As usual, this turned out to be quite a
bit harder to implement than it initially seemed.) This isn't really
intended to be widely used, though it is occasionally convenient for
type hinting in complex expressions. However, it may be useful for
implementing certain aspects of full let-polymorphism.