fix: further improve ∀ unification
A bunch of cases with ∀ types were still being handled badly. The worst
case was unifying a type variable like `X` with a bound type like
`∀<X,Y>`. I'd previously thought that the mgu should just be the latter.
But this is clearly wrong when considering pair like <X,X> vs
<∀<X,Y>,∀<X,Y>>, where returning the latter fails to capture the
identity constraint between the input and output type. A type like
∀<<X,Y>,<X,Y>> might be better, but then this will fail to capture the
mapping of X to a type. At the level of an individual unification this
might be ok, but for unify calls where X may appear elsewhere in the
global context (and the updated assignment is used) it is not. I'm
pretty sure that the only option in the present type system is to move
to global fresh types, and return a principal like <<?1,?2>,<?1,?2>>.
There are lots of other cases that trigger something like this scenario,
and I have rewritten ∀ unification once again to handle them. Also,
versions of this scenario required a debind call in VariableType
unification; this is unfortunate as it leads to mappings that don't get
cleaned up, but I don't see a way around it right now.