feat: named let-polymorphism via assignments
This commit uses ∀-bound types in assignments to signal that a variable
name should be taken as let-polymorphic (previously only anonymous
let-polymorphism was allowed). This part was easy, but what was quite
involved was doing the TypeEnv manipulation involved in this change, in
order to coherently track and resolve specializations. There are
probably still issues as of this commit. For the most part, if ∀ types
aren't used, then things shouldn't get too strange.

Some weirdnesses that I know of:

* It is now possible to write functions whose type constraints on
  application are not fully represented in the type signature, by using
  ∀-bound types as arguments. I'm not entirely sure whether this is a
  good thing or not, but it was essentially entailed by the general
  implementation.
* There are (buggy?) cases where a primary polymorphic type can be
  annotated, which allows inferring an expression that fails repr-parse.
  This needs something more, maybe an explicit Let operator.

Subsidiary changes:

* regularize_type_env needed to be called on every combination. Maybe
  now overkill as-is?
* I needed a good alpha equivalence function for type variables, and the
  existing one wasn't ideal. So this commit contains a rewritten one
  that is not much more costly than equality checking.