fix: prevent ∀-based let-polymorphism from bottom up type hints
As of 6fc3d2cd4 there is very general code to handle let-polymorphism if
terms are annotated with ∀ types. Because it is so general, it can cope
with these type hints being written on individual terms (with some
caveats). However, the result is quite weird from a metalanguage
perspective, in that you get an expression whose repr does not represent
in any way the let-polymorphism, and so the equality that te(repr(te(x))
== te(x) will fail (because the repr will be parsed without
let-polymorphism). The situation is different for functions whose bound
variable indicates the type, and for let-polymorphism via assignment,
since in both of these cases there is something providing top-down
guidance -- and further, we don't care about the repr-parse equality for
assignment substitutions. I considered more elaborate solutions, such as
adding some kind of explicit `Let` operator. But the simplest fix for
now is to simply only allow the latter two cases (which both work via
assignment anyways). Let-polymorphism via type hinting is probably
pretty confusing to most people anyways.